requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
anthropic>=0.40.0
pdfplumber==0.11.1
bcrypt==4.1.3
python-dotenv==1.0.1
flask-cors==4.0.1
